third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-inherited-variable.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#parent {
    --foo: {
        background-color: green;
        width: 100px;
        height: 200px;
    };
}
#child {
    @apply --foo;
}
</style>
<div id=parent>
    <div id=child>
    </div>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(child).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(child).width, "100px");
    assert_equals(getComputedStyle(child).height, "200px");
}, "@apply when the variable is inherited");
</script>
